+++ builtwithdjango/logging_utils.py
@@ -1,0 +1,6 @@
+import logfire
+
+
+def scrubbing_callback(m: logfire.ScrubMatch):
+    if m.path == ("attributes", "cookies"):
+        return m.value