+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project tries to adhere to [Semantic Versioning](https://semver.org/spe
 - Support for getting ip address of user and passing it to buttondown
 - Support for logfire
 
+### Removed
+- Usage of opentelemetry directly
 
 ## [0.4.5] - 2025-02-13
 ### Added